requirements.txt
transformers
hydra-core
hydra_colorlog
omegaconf
torch
pandas
seaborn
matplotlib
datasets
accelerate
triton
triton_kernels
tensorboard
tqdm
scikit-learn
nltk
hydra-optuna-sweeper